+++ app/Models/BackendCms/PermissionUser.php
@@ -115,22 +115,23 @@ class PermissionUser extends BaseModel
     public function updatePermissUser($arrPermissForm = [], $user_id = 0, $project_code = 0)
     {
         $edit = false;
-        if (empty($arrPermissForm) || $user_id <= 0 || $project_code <= 0)
+        if ($user_id <= 0 || $project_code <= 0)
             return $edit;
 
-        foreach ($arrPermissForm as $permiss_code => $arrMenu) {
-            PermissionUser::where('user_id', $user_id)
-                ->where('project_code', $project_code)->delete();
-        }
-        foreach ($arrPermissForm as $permiss_code => $arrMenu) {
-            foreach ($arrMenu as $k => $menu_id) {
-                $arrInsert['user_id'] = $user_id;
-                $arrInsert['project_code'] = $project_code;
-                $arrInsert['menu_id'] = $menu_id;
-                $arrInsert['permiss_code'] = $permiss_code;
-                $arrInsert['is_active'] = STATUS_INT_MOT;
-                if ($this->editItem($arrInsert)) {
-                    $edit = true;
+        PermissionUser::where('user_id', $user_id)->where('project_code', $project_code)->delete();
+        $edit = true;
+
+        if(!empty($arrPermissForm)){
+            foreach ($arrPermissForm as $permiss_code => $arrMenu) {
+                foreach ($arrMenu as $k => $menu_id) {
+                    $arrInsert['user_id'] = $user_id;
+                    $arrInsert['project_code'] = $project_code;
+                    $arrInsert['menu_id'] = $menu_id;
+                    $arrInsert['permiss_code'] = $permiss_code;
+                    $arrInsert['is_active'] = STATUS_INT_MOT;
+                    if ($this->editItem($arrInsert)) {
+                        $edit = true;
+                    }
                 }
             }
         }
